frontend/requirements.txt
# frontend/requirements.txt
Flask
requests
firebase-admin
pillow

